Refactor state variables for better performance
I found there was unnessary rerendering due to a state variable that
tracks the number of images that have loaded. As this updated with
every load the whole page rerendered unnessarily. Thus this count
was managed by a Ref and the state was only updated when all images
were loaded.